+++ src/package/themes/default/components/spinner.ts
@@ -1,0 +1,36 @@
+export default {
+    size: '1.5rem',
+    thickness: '2px',
+    color: 'var(--vf-blue-600)',
+    trackColor: 'rgba(var(--vf-blue-600-rgb), 0.2)',
+    gap: '0.5rem',
+    labelColor: 'var(--vf-secondary-text-color)',
+    labelFontSize: '0.875rem',
+    animationDuration: '0.8s',
+    overlayMinHeight: '4rem',
+    overlayPadding: '0.75rem',
+    overlayBorderRadius: 'var(--vf-radii-md)',
+    overlayBackgroundColor: 'var(--vf-bg-soft-color)',
+    info: {
+        color: 'var(--vf-sky-600)',
+    },
+    success: {
+        color: 'var(--vf-green-600)',
+    },
+    warn: {
+        color: 'var(--vf-yellow-600)',
+    },
+    danger: {
+        color: 'var(--vf-red-600)',
+    },
+    small: {
+        size: '1.1rem',
+        thickness: '2px',
+        labelFontSize: '0.8rem',
+    },
+    large: {
+        size: '2rem',
+        thickness: '3px',
+        labelFontSize: '1rem',
+    },
+};